Major Deprecation of old processing config
-deprecate old create_run_config from ConfigCreator
-depreacate helpers.py, replace with helpers_new.py in pipelines
-move decimation_level_config.py into sandbox, to support Gary's matlab z-files if needed
-add exceptions in transfer_function_collection if old processing config detected
-remove old processing_config.py

[Issue(s): #178]
