+++ aurora/transfer_function/transfer_function_collection.py
@@ -363,7 +363,9 @@ class TransferFunctionCollection(object):
                 dec_level_config = tf.processing_config.decimations[i_dec]
             except:
                 print("OLD STYLE ProceessingConfig in use")
+                print("This is supposed to be deprecated in 0.0.3")
                 dec_level_config = tf.processing_config
+                raise Exception
 
             for band in tf.frequency_bands.bands(direction="increasing_period"):
                 #print(f"band {band}")
@@ -374,8 +376,11 @@ class TransferFunctionCollection(object):
                     sample_rate = tf.processing_config.sample_rate
                     num_samples_window = tf.processing_config.num_samples_window
                 except AttributeError:
+                    print("OLD STYLE ProceessingConfig in use")
+                    print("This is supposed to be deprecated in 0.0.3")
                     sample_rate = dec_level_config.decimation.sample_rate
                     num_samples_window = dec_level_config.window.num_samples
+                    raise Exception
                 freqs = np.fft.fftfreq(num_samples_window, 1.0 / sample_rate)
                 fc_indices = band.fourier_coefficient_indices(freqs)
                 # </Make a method of processing config?>
